+++ atelier-guild-rank/src/shared/services/game-flow/game-flow-manager.interface.ts
@@ -1,0 +1,178 @@
+/**
+ * game-flow-manager.interface.ts - GameFlowManagerインターフェース
+ *
+ * TASK-0017: GameFlowManager実装
+ *
+ * @description
+ * ゲーム全体の進行を統括する中心的なサービスのインターフェース。
+ * ゲームの開始・終了、日の進行、フェーズ遷移、ゲーム終了判定を提供する。
+ *
+ * @信頼性レベル 🔵
+ * - 設計文書・要件定義書に基づいた定義
+ * - ゲームフロー全体の統括機能
+ */
+
+import type { GamePhase, GuildRank, ISaveData } from '@shared/types';
+
+// =============================================================================
+// ゲーム終了条件
+// =============================================================================
+
+/**
+ * 【機能概要】: ゲーム終了条件
+ * 【実装方針】: ゲームオーバー・ゲームクリア時の情報を格納
+ * 🔵 信頼性レベル: 設計文書に明記
+ */
+export interface GameEndCondition {
+  /** 終了タイプ */
+  type: 'game_over' | 'game_clear';
+  /** 終了理由 */
+  reason: string;
+  /** 最終ランク */
+  finalRank: GuildRank;
+  /** 総日数 */
+  totalDays: number;
+}
+
+// =============================================================================
+// GameFlowManagerインターフェース
+// =============================================================================
+
+/**
+ * 【機能概要】: GameFlowManagerインターフェース
+ * 【実装方針】: ゲーム開始、日の進行、フェーズ遷移、ゲーム終了判定を提供
+ * 🔵 信頼性レベル: 設計文書・要件定義書に明記
+ */
+export interface IGameFlowManager {
+  // =============================================================================
+  // ゲーム開始
+  // =============================================================================
+
+  /**
+   * 【機能概要】: 新規ゲームを開始
+   * 【実装方針】: StateManagerとDeckServiceを初期化し、startDay()を呼び出す
+   * 【テスト対応】: T-0017-01 新規ゲーム開始
+   * 🔵 信頼性レベル: 設計文書に明記
+   */
+  startNewGame(): void;
+
+  /**
+   * 【機能概要】: セーブデータからゲームを再開
+   * 【実装方針】: セーブデータから各サービスの状態を復元
+   * 【例外】: ApplicationError(ErrorCodes.INVALID_SAVE_DATA) - セーブデータが不正な場合
+   * 🟡 信頼性レベル: 要件定義書から推測
+   *
+   * @param saveData - セーブデータ
+   */
+  continueGame(saveData: ISaveData): void;
+
+  // =============================================================================
+  // 日の進行
+  // =============================================================================
+
+  /**
+   * 【機能概要】: 日を開始
+   * 【実装方針】: AP回復→依頼生成→DAY_STARTEDイベント発行→QUEST_ACCEPTフェーズに遷移
+   * 【テスト対応】: T-0017-02 日開始処理
+   * 🔵 信頼性レベル: 設計文書に明記
+   */
+  startDay(): void;
+
+  /**
+   * 【機能概要】: 日を終了
+   * 【実装方針】: 期限処理→日数更新→DAY_ENDEDイベント発行→ゲーム終了判定→次の日へ or ゲーム終了
+   * 【テスト対応】: T-0017-05 日終了処理
+   * 🔵 信頼性レベル: 設計文書に明記
+   */
+  endDay(): void;
+
+  // =============================================================================
+  // フェーズ進行
+  // =============================================================================
+
+  /**
+   * 【機能概要】: 指定されたフェーズに遷移
+   * 【実装方針】: StateManager.setPhase()を呼び出す
+   * 【例外】: ApplicationError(ErrorCodes.INVALID_PHASE_TRANSITION) - 無効なフェーズ遷移の場合
+   * 【テスト対応】: T-0017-03 フェーズ進行
+   * 🔵 信頼性レベル: 設計文書に明記
+   *
+   * @param phase - 遷移先のフェーズ
+   */
+  startPhase(phase: GamePhase): void;
+
+  /**
+   * 【機能概要】: 現在のフェーズを終了し、次のフェーズに遷移
+   * 【実装方針】: QUEST_ACCEPT→GATHERING→ALCHEMY→DELIVERY→endDay()
+   * 【テスト対応】: T-0017-04 endPhase()で次のフェーズに遷移
+   * 🟡 信頼性レベル: 要件定義書から推測
+   */
+  endPhase(): void;
+
+  /**
+   * 【機能概要】: 現在のフェーズをスキップして次のフェーズに遷移
+   * 【実装方針】: endPhase()と同様だが、スキップとして記録
+   * 【テスト対応】: T-0017-09 skipPhase()でフェーズをスキップ
+   * 🟡 信頼性レベル: 要件定義書から推測
+   */
+  skipPhase(): void;
+
+  // =============================================================================
+  // ゲーム終了判定
+  // =============================================================================
+
+  /**
+   * 【機能概要】: ゲームオーバー判定
+   * 【実装方針】: remainingDays <= 0 && currentRank !== GuildRank.S の場合にGameEndConditionを返す
+   * 【テスト対応】: T-0017-B01 残り日数が0でSランク未到達の場合、ゲームオーバー判定
+   * 🔵 信頼性レベル: 設計文書に明記
+   *
+   * @returns ゲームオーバーの場合GameEndCondition、継続の場合null
+   */
+  checkGameOver(): GameEndCondition | null;
+
+  /**
+   * 【機能概要】: ゲームクリア判定
+   * 【実装方針】: currentRank === GuildRank.S の場合にGameEndConditionを返す
+   * 【テスト対応】: T-0017-06 ゲームクリア条件の判定
+   * 🔵 信頼性レベル: 設計文書に明記
+   *
+   * @returns ゲームクリアの場合GameEndCondition、未到達の場合null
+   */
+  checkGameClear(): GameEndCondition | null;
+
+  // =============================================================================
+  // アクション
+  // =============================================================================
+
+  /**
+   * 【機能概要】: 休憩アクション（AP消費なしで日を進める）
+   * 【実装方針】: 手札を2枚捨てて2枚ドロー→endDay()を呼び出す
+   * 【テスト対応】: T-0017-10 rest()でAP消費なしで日が進む
+   * 🟡 信頼性レベル: タスク定義から推測
+   */
+  rest(): void;
+
+  // =============================================================================
+  // 状態取得
+  // =============================================================================
+
+  /**
+   * 【機能概要】: 現在のフェーズを取得
+   * 【実装方針】: StateManager.getState().currentPhaseを返す
+   * 【テスト対応】: T-0017-08 getCurrentPhase()で現在のフェーズを取得
+   * 🟡 信頼性レベル: 要件定義書から推測
+   *
+   * @returns 現在のフェーズ
+   */
+  getCurrentPhase(): GamePhase;
+
+  /**
+   * 【機能概要】: 次のフェーズに進めるかを判定
+   * 【実装方針】: 現在のフェーズの必須アクションが完了しているかをチェック
+   * 🟡 信頼性レベル: 要件定義書から推測
+   *
+   * @returns 進める場合true
+   */
+  canAdvancePhase(): boolean;
+}
