+++ atelier-guild-rank/src/shared/services/save-load/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Shared Services Save/Load
+ * SaveLoadServiceの公開エクスポート
+ *
+ * TASK-0104: application層からshared/servicesに移動
+ */
+
+export type { ISaveLoadService } from './save-load-service';
+export { SaveLoadService } from './save-load-service';
